refactored token type to be data oriented

+++ error/error.go
@@ -14,7 +14,7 @@ func ScanPanic(path *string, line int, reason string) {
 	panic(msg)
 }
 
-func ParsePanic(path *string, line int, reason string) {
+func ParsePanic(path *string, line uint16, reason string) {
 	msg := fmt.Sprintf("%s:%d: %s", *path, line, reason)
 	panic(msg)
 }
